+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
+## [6.1.0] - 2022-12-10
+
 ### Changed
 - Enhance plotting with custom `x`- and `y`-data in `CharacteristicCurve.plot(x, y)` and allow a list of items for force evaluation in `CharacteristicCurve(items=[...])` to be passed.
 - Enhance `math.linsteps(points=[0, 5, 0], num=[5, 10])` by supporting a list of substeps.